% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcharter_facet.R
\name{runcharter_facet}
\alias{runcharter_facet}
\title{Create run chart with highlighted improvements where applicable}
\usage{
runcharter_facet(df, med_rows = 13, runlength = 9,
  chart_title = NULL, chart_subtitle = NULL, direction = "below",
  faceted = TRUE, n_facet_cols = NULL, ...)
}
\arguments{
\item{df}{dataframe containing columns "date", "y", "grp"}

\item{med_rows}{the number of rows you wish to base the initial median calculation over}

\item{runlength}{how long a run of consecutive points should be before re-basing the median}

\item{chart_title}{title for the  final chart}

\item{chart_subtitle}{subtitle for chart}

\item{direction}{look for runs "below" or "above" the median}

\item{faceted}{whether the plot is faceted or not - must be TRUE to call this function}

\item{n_facet_cols}{the number of columns to facet by}

\item{...}{further arguments passed on to function}
}
\value{
runchart and a dataframe showing sustained run data if appropriate
}
\description{
This will plot the original dataframe, with highlighted runs of improvement.
It will also return a dataframe showing the improvment data
}
\examples{
\donttest{
runcharter_facet(data, med_rows = 13, runlength = 9, chart_title = "Automated runs analysis",
direction = "above")
}


}
