% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signals}
\alias{signals}
\title{#' 220 grouped observations over time.}
\format{
A data frame with 220 rows and 4 variables:
\describe{
  \item{grp}{a grouping variable, representing a specific department}
  \item{y}{integers representing counts of an event over time}
  \item{date}{date of the observation, by month}
}
}
\usage{
signals
}
\description{
A dataset containing four equal groups of 55 integers
simulating signals of improvement in multiple directions
relative to their respective baseline medians.
}
\keyword{datasets}
