% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcharter.R
\name{runcharter}
\alias{runcharter}
\title{Create run chart with highlighted improvements where applicable}
\usage{
runcharter(df, med_rows = 13, runlength = 9, chart_title = NULL,
  chart_subtitle = NULL, direction = "below", faceted = FALSE,
  n_facet_cols = NULL, ...)
}
\arguments{
\item{df}{dataframe containing columns "date", "y", "grp"}

\item{med_rows}{the number of rows you wish to base the initial median calculation over}

\item{runlength}{how long a run of consecutive points should be before re-basing the median}

\item{chart_title}{title for the  final chart}

\item{chart_subtitle}{subtitle for chart}

\item{direction}{look for runs "below" or "above" the median}

\item{faceted}{if you want to plot multiple plots at once in a faceted / trellis display set this to TRUE}

\item{n_facet_cols}{the number of columns required for a faceted plot. Ignored if faceted is set to FALSE}

\item{...}{further arguments passed on to function}
}
\value{
runchart and a dataframe showing sustained run data if appropriate
}
\description{
This will plot the original dataframe, with highlighted runs of improvement.
It will also return a dataframe showing the improvment data
}
\examples{
\donttest{
runcharter(data, med_rows = 13, runlength = 9, chart_title = "Automated runs analysis",
direction = "above", faceted = FALSE)
}


}
