% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runcharter.R
\name{runcharter}
\alias{runcharter}
\title{Create run chart with highlighted improvements where applicable}
\usage{
runcharter(df, med_rows = 13, runlength = 9, chart_title = NULL,
  chart_subtitle = NULL, direction = "below", faceted = TRUE,
  facet_cols = NULL, save_plot = FALSE, plot_extension = "png",
  verbose = FALSE, ...)
}
\arguments{
\item{df}{dataframe containing columns "date", "y", "grp"}

\item{med_rows}{number of rows to base the initial median calculation over}

\item{runlength}{how long a run of consecutive points should be before re-basing the median}

\item{chart_title}{title for the  final chart}

\item{chart_subtitle}{subtitle for chart}

\item{direction}{look for runs "below" or "above" the median, or "both"}

\item{faceted}{if you  dont need a faceted / trellis display set this to FALSE}

\item{facet_cols}{the number of columns required for a faceted plot. Ignored if faceted is set to FALSE}

\item{save_plot}{should the plot be saved?  Calls ggsave on the last plot, saving in the current directory, if TRUE.}

\item{plot_extension}{one of "png","pdf" or other valid extension for saving ggplot2 plots. Used in the call to ggsave.}

\item{...}{further arguments passed on to function}
}
\value{
run chart(s) and a dataframe showing sustained run data if appropriate
}
\description{
This will plot the original dataframe, with highlighted runs of improvement.
It will also return a dataframe showing the improvment data
}
\examples{
\donttest{
runcharter(signals, med_rows = 13, runlength = 9,
chart_title = "Automated runs analysis",
direction = "above", faceted = TRUE,
facet_cols = 2, save_plot = TRUE, plot_extension = "png")
}


}
